[BoundsSafety] Allow runtime calls in attr args for attribute-only mode
For -fbounds-safety, function calls in e.g. the counted_by argument are
only allowed if the function can be evaluated to a constant value at
compile time. This is because of the complexity of preventing a function
return value which depends on a runtime variable from changing, i.e. it
would be really hard to make that safe.

With -fexperimental-bounds-safety-attributes the same attributes are not
enforced however, but may instead serve as unchecked assertions of fact,
to be consumed by programmers and tooling alike. This allows annotations
for functions like this:

```
char * __counted_by(strlen(s)) str_to_buf(char * __null_terminated s) {
  return s;
}
```

-fbounds-safety cannot verify that the function above is safe, but in
less strict models it can nonetheless be useful information to
communicate.